Update user statistics UI
